Move the IStringLog and its implementations from Services to Logging.
